+++ .changeset/short-windows-buy.md
@@ -1,0 +1,12 @@
+---
+"@evmts/bundler": patch
+"@evmts/esbuild-plugin": patch
+"@evmts/webpack-plugin": patch
+"@evmts/rollup-plugin": patch
+"@evmts/rspack-plugin": patch
+"@evmts/vite-plugin": patch
+"@evmts/bun-plugin": patch
+"@evmts/ts-plugin": patch
+---
+
+Reexported @evmts/config from bundlers
