+++ bundlers/bundler/src/index.ts
@@ -3,3 +3,4 @@ export * from './unplugin'
 export * from './types'
 export * from './bundler'
 export * from './runtime'
+export * from '@evmts/config'
