+++ bundlers/esbuild/package.json
@@ -44,7 +44,8 @@
     "test:ui": "vitest --ui"
   },
   "dependencies": {
-    "@evmts/bundler": "workspace:^"
+    "@evmts/bundler": "workspace:^",
+    "@evmts/config": "workspace:^"
   },
   "devDependencies": {
     "@evmts/tsconfig": "workspace:^",
