+++ bundlers/esbuild/src/index.ts
@@ -1,1 +1,2 @@
 export { esbuildPluginEvmts } from '@evmts/bundler'
+export * from '@evmts/config'
