+++ bundlers/rollup/src/index.ts
@@ -1,1 +1,2 @@
 export { rollupPluginEvmts } from '@evmts/bundler'
+export * from '@evmts/config'
