+++ bundlers/rspack/src/index.ts
@@ -1,1 +1,2 @@
 export { rspackPluginEvmts } from '@evmts/bundler'
+export * from '@evmts/config'
