+++ bundlers/vite/src/index.ts
@@ -1,1 +1,2 @@
 export { vitePluginEvmts } from '@evmts/bundler'
+export * from '@evmts/config'
