+++ bundlers/webpack/src/index.ts
@@ -1,1 +1,2 @@
 export { webpackPluginEvmts } from '@evmts/bundler'
+export * from '@evmts/config'
